refactor: move rails on artist artwork tab to overview (about) tab
