+++ CHANGELOG.yml
@@ -6,6 +6,8 @@ upcoming:
   user_facing:
     - Combine gallery and institution artwork filters - mdole
     - Allows multi-select filters to be searchable - damon
+    - Rename artist about tab to artist overview tab - iskounen
+    - Move rails on artist artwork tab to artist overview tab - iskounen
 
 releases:
   - version: 6.8.4
